refactor(query): extract query pipeline mutation logic to reduce complexity
Moves string and parsing mutation steps (e.g. padding, boosting syntax, parsing fields) into a separate \query_pipeline.py\ to simplify \query.py\. \query.py\ now focuses primarily on Elasticsearch mapping and orchestration.
